Refactor versioning approach
